pyproject.toml
[project]
name = "gps-genealogy-agents"
version = "0.3.1"
description = "Multi-agent AI genealogical research system using Semantic Kernel and AutoGen, meeting GPS (Genealogical Proof Standard) with Gramps integration"
readme = "README.md"
requires-python = ">=3.13"
license = { text = "MIT" }
authors = [{ name = "Thomas Vincent", email = "thomasvincent@gmail.com" }]
keywords = [
    "genealogy",
    "ai",
    "agents",
    "autogen",
    "semantic-kernel",
    "gps",
    "gramps",
    "multi-agent",
    "agentic-ai",
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Science/Research",
    "Intended Audience :: End Users/Desktop",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3.13",
    "Programming Language :: Python :: 3.14",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "Topic :: Sociology :: Genealogy",
]

dependencies = [
    # Semantic Kernel + AutoGen multi-agent framework
    "semantic-kernel>=1.39.0",
    "autogen-ext[anthropic,openai]>=0.7.5",
    "tenacity>=8.2.0",
    # LLM providers
    "anthropic>=0.40.0",
    "openai>=1.0.0",
    "langchain-anthropic>=0.2.0",
    "langchain-openai>=0.2.0",
    # Gramps integration
    "gramps>=5.2.0,<7.0.0",
    "orjson>=3.9.0",
    "lxml>=5.0.0",
    # Core utilities
    "pydantic>=2.5.0",
    "pydantic-settings>=2.1.0",
    "aiosqlite>=0.20.0",
    "httpx>=0.27.0",
    "beautifulsoup4>=4.12.0",
    # Logging
    "structlog>=24.1.0",
    # CLI and formatting
    "typer>=0.12.0",
    "rich>=13.7.0",
    # Templates and utilities
    "jinja2>=3.1.0",
    "python-dateutil>=2.8.0",
    "pyyaml>=6.0.0",
    "python-dotenv>=1.0.0",
    # UUID7 support (built-in Python 3.14+, polyfill for 3.13)
    "uuid-utils>=0.9.0",
    "werkzeug>=3.1.5",
    "autogen-agentchat>=0.4.0",
    "jsonschema>=4.20.0",
]

[project.optional-dependencies]
# ChromaDB for semantic memory (requires Python <3.14 due to onnxruntime)
chromadb = [
    "chromadb>=1.0.0",
]

# RocksDB for immutable fact ledger
rocksdb = [
    "python-rocksdb>=0.7.0",
]

# Gramps GUI plugin dependencies (for desktop Gramps)
gramps-gui = [
    "PyGObject>=3.42.0",
    "pycairo>=1.20.0",
]

# Full Gramps integration with all features
gramps-full = [
    "PyGObject>=3.42.0",
    "pycairo>=1.20.0",
    "Pillow>=10.0.0",
    "PyICU>=2.10.0",
]

# Web API dependencies
web = [
    "fastapi>=0.109.0",
    "uvicorn>=0.27.0",
    "python-multipart>=0.0.20",
]

# Additional LLM providers
azure = [
    "azure-identity>=1.15.0",
]
ollama = [
    "ollama>=0.1.0",
]

# Development dependencies
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "pytest-cov>=4.1.0",
    "hypothesis>=6.100.0",  # Property-based testing for edge cases
    "ruff>=0.8.0",  # Best-in-class linter/formatter (replaces flake8, black, isort)
    "mypy>=1.10.0",
    "pre-commit>=3.6.0",
    "types-PyYAML>=6.0.0",
    "types-python-dateutil>=2.8.0",
    "types-beautifulsoup4>=4.12.0",
]

# All optional dependencies (excludes chromadb which requires Python <3.14)
all = [
    "gps-genealogy-agents[rocksdb,gramps-full,web,dev]",
]

[project.scripts]
gps-agents = "gps_agents.cli:app"
genai = "gps_agents.cli:app"

[project.urls]
Homepage = "https://github.com/thomasvincent/gps-genealogy-agents"
Documentation = "https://github.com/thomasvincent/gps-genealogy-agents#readme"
Repository = "https://github.com/thomasvincent/gps-genealogy-agents"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/gps_agents"]

[tool.ruff]
line-length = 100
target-version = "py313"
src = ["src", "tests"]

[tool.ruff.lint]
select = [
    "E",      # pycodestyle errors
    "W",      # pycodestyle warnings
    "F",      # Pyflakes
    "I",      # isort
    "N",      # pep8-naming
    "B",      # flake8-bugbear
    "C4",     # flake8-comprehensions
    "UP",     # pyupgrade
    "ARG",    # flake8-unused-arguments
    "SIM",    # flake8-simplify
    "S",      # flake8-bandit (security)
    "A",      # flake8-builtins
    "COM",    # flake8-commas
    "DTZ",    # flake8-datetimez
    "T10",    # flake8-debugger
    "ISC",    # flake8-implicit-str-concat
    "ICN",    # flake8-import-conventions
    "PIE",    # flake8-pie
    "PT",     # flake8-pytest-style
    "Q",      # flake8-quotes
    "RSE",    # flake8-raise
    "RET",    # flake8-return
    "SLF",    # flake8-self
    "TCH",    # flake8-type-checking
    "RUF",    # Ruff-specific rules
    "PERF",   # Perflint
    "FURB",   # refurb
    "LOG",    # flake8-logging
    "TRY",    # tryceratops
]
ignore = [
    "E501",    # line length handled by formatter
    "S101",    # assert is fine in tests
    "COM812",  # trailing comma conflicts with formatter
    "ISC001",  # conflicts with formatter
    "TRY003",  # long messages in exceptions are fine
    "PERF401", # list comprehension in loop - often less readable
    "S608",    # SQL injection - we use parameterized queries
    "S112",    # try-except-continue - acceptable in some cases
    "RUF001",  # ambiguous unicode - en-dash is intentional
    "TC001",   # Pydantic needs runtime imports, not TYPE_CHECKING
    "TC002",   # same as TC001
]

[tool.ruff.lint.per-file-ignores]
"tests/**/*.py" = [
    "S101",   # assert is expected in tests
    "ARG",    # unused arguments in fixtures
    "SLF001", # private member access in tests
]

[tool.ruff.lint.isort]
known-first-party = ["gps_agents"]
required-imports = ["from __future__ import annotations"]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
skip-magic-trailing-comma = false
line-ending = "auto"
docstring-code-format = true

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
addopts = "-v --tb=short"
filterwarnings = [
    "ignore::DeprecationWarning:datetime",
    "ignore::pydantic.warnings.PydanticDeprecatedSince20",
]

[tool.hypothesis]
# Property-based testing configuration
max_examples = 100
deadline = 5000  # 5 seconds
suppress_health_check = ["too_slow"]

[tool.mypy]
python_version = "3.13"
strict = true
warn_return_any = true
warn_unused_configs = true
ignore_missing_imports = true
# Per-module overrides for gradual typing
[[tool.mypy.overrides]]
module = [
    "gps_agents.graph",
    "gps_agents.cli",
    "gps_agents.autogen.*",
    "gps_agents.sk.*",
    "gps_agents.sources.*",
]
# Allow less strict typing for complex modules being gradually typed
disallow_untyped_defs = false
disallow_untyped_calls = false
